requirements.txt
geopy
telebot
pytz
python-dotenv